Move fallback class into framework.time package
