+++ src/main/java/org/opentripplanner/apis/gtfs/mapping/RouteRequestMapper.java
@@ -15,9 +15,9 @@ import javax.annotation.Nonnull;
 import org.opentripplanner.api.common.LocationStringParser;
 import org.opentripplanner.api.parameter.QualifiedMode;
 import org.opentripplanner.api.parameter.QualifiedModeSet;
-import org.opentripplanner.api.support.ZoneIdFallback;
 import org.opentripplanner.apis.gtfs.GraphQLRequestContext;
 import org.opentripplanner.framework.graphql.GraphQLUtils;
+import org.opentripplanner.framework.time.ZoneIdFallback;
 import org.opentripplanner.model.GenericLocation;
 import org.opentripplanner.routing.api.request.RouteRequest;
 import org.opentripplanner.routing.api.request.framework.CostLinearFunction;
