+++ src/main/java/org/opentripplanner/framework/time/ZoneIdFallback.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.api.support;
+package org.opentripplanner.framework.time;
 
 import java.time.ZoneId;
 import javax.annotation.Nullable;
